Separate out utility functions and clean up file.
